+++ src/test/scala/com/jcdecaux/setl/storage/repository/streaming/StreamingRepositorySuite.scala
@@ -3,6 +3,7 @@ package com.jcdecaux.setl.storage.repository.streaming
 import com.jcdecaux.setl.SparkSessionBuilder
 import com.jcdecaux.setl.exception.InvalidConnectorException
 import com.jcdecaux.setl.storage.connector.CSVConnector
+import com.jcdecaux.setl.storage.repository.SparkRepository
 import org.apache.spark.sql.SparkSession
 import org.scalatest.funsuite.AnyFunSuite
 
@@ -18,10 +19,10 @@ class StreamingRepositorySuite extends AnyFunSuite {
       "header" -> "true"
     )
     val csvConnector = new CSVConnector(csvOutputConf)
-    val repo = new StreamingRepository[TestClass]()
+    val repo = new SparkRepository[TestClass]().setConnector(csvConnector)
 
-    assertThrows[InvalidConnectorException](repo.setConnector(csvConnector))
-
+    assertThrows[InvalidConnectorException](repo.awaitTermination())
+    assertThrows[InvalidConnectorException](repo.awaitTerminationOrTimeout(1))
   }
 
 }